Feature flag removal, SequencescapeTubes,
Feature flag removed for Sequencescapr V1
SequencescapeTubes added
Common functions moved to sequencescapeUtils.js
