+++ src/views/GeneralReception.vue
@@ -189,8 +189,6 @@ import TractionHeading from '../components/TractionHeading.vue'
 import LibraryTypeSelect from '@/components/shared/LibraryTypeSelect'
 import DataTypeSelect from '@/components/shared/DataTypeSelect'
 import { defaultRequestOptions } from '@/lib/receptions'
-// DEPRECATE-DPL-877 when is enabled by default, remove this
-import useSWRV from 'swrv'
 
 // We don't expect the modal to display without a message. If we end up in this
 // state then something has gone horribly wrong.
@@ -225,7 +223,7 @@ export default {
     modalState: defaultModal(),
   }),
   computed: {
-    reception: ({ receptions, source, version }) => receptions[source][version],
+    reception: ({ receptions, source }) => receptions[source],
     api() {
       return this.$store.getters.api
     }, // can't use this in arrow function
@@ -235,13 +233,6 @@ export default {
     barcodeCount: ({ barcodeArray }) => barcodeArray.length,
     presentRequestOptions: ({ requestOptions }) =>
       Object.fromEntries(Object.entries(requestOptions).filter(([, v]) => v)),
-    // DEPRECATE-DPL-877 once is enabled by default, remove this
-    version: () => {
-      // we only care if the feature is enabled or not, so we can just check for the presence of the feature
-      // if the feature flag can't be returned for any reason, we'll default to v1
-      const { data } = useSWRV(`${import.meta.env.VITE_TRACTION_BASE_URL}/flipper/api/actors/User`)
-      return data?.value?.features['dpl_877_reception_request']?.enabled ? 'v2' : 'v1'
-    },
   },
   methods: {
     importStarted({ message }) {
@@ -266,27 +257,16 @@ export default {
           foundBarcodes,
           attributes,
         )
+        const messages = createMessages({
+          barcodes: this.barcodeArray,
+          response,
+          reception: this.reception,
+        })
 
-        // v2 returns a different response to v1, so we need to handle it differently
-        // DEPRECATE-DPL-877 once enabled by default, remove this line
-        if (this.version === 'v2') {
-          const messages = createMessages({
-            barcodes: this.barcodeArray,
-            response,
-            reception: this.reception,
-          })
-
-          // we create a different alert for each message
-          messages.forEach(({ type, text }) => {
-            this.showAlert(text, type)
-          })
-          // DEPRECATE-DPL-877 once is enabled by default, remove this
-        } else {
-          this.showAlert(
-            `Imported ${foundBarcodes.size} labware(s) from ${this.reception.text}`,
-            'success',
-          )
-        }
+        // we create a different alert for each message
+        messages.forEach(({ type, text }) => {
+          this.showAlert(text, type)
+        })
       } catch (e) {
         console.error(e)
         this.showAlert(e, 'danger')
