+++ tests/e2e/specs/reception/import_samples_from_samples_extraction.cy.js
@@ -10,31 +10,8 @@ describe('Import samples from Samples extraction, for Pacbio', () => {
     })
   })
 
-  // DEPRECATE-DPL-877 once enabled by default, remove this
-  it('Successfully - v1', () => {
-    cy.withFlags({
-      dpl_877_reception_request: { enabled: false },
-    })
-    cy.visit('#/reception')
-    cy.get('[data-type="source-list"]').select('Samples Extraction')
-    cy.contains('Scan barcodes')
-    cy.get('#barcodes').type('SE108532I')
-    cy.intercept('/api/v1/assets?filter[barcode]=SE108532I', {
-      fixture: 'sampleExtractionTubesWithSample.json',
-    })
-    cy.intercept('POST', '/v1/receptions', { fixture: 'tractionPacbioRequest.json' }).as(
-      'postPayload',
-    )
-    cy.contains('Import 1 labware into PacBio from Samples Extraction')
-    cy.get('[data-action="import-labware"]').click()
-    cy.contains('Imported 1 labware(s) from Samples Extraction')
-  })
-
+  
   it('Successfully - v2', () => {
-    // DEPRECATE-DPL-877 once enabled by default, remove this
-    cy.withFlags({
-      dpl_877_reception_request: { enabled: true },
-    })
     cy.visit('#/reception')
     cy.get('[data-type="source-list"]').select('Samples Extraction')
     cy.contains('Scan barcodes')
@@ -59,10 +36,6 @@ describe('Import samples from Samples extraction, for Pacbio', () => {
   })
 
   it('Unsuccessfully - When tubes are missing', () => {
-    // DEPRECATE-DPL-877 once enabled by default, remove this line
-    cy.withFlags({
-      dpl_877_reception_request: { enabled: false },
-    })
     cy.visit('#/reception')
     cy.get('[data-type="source-list"]').select('Samples Extraction')
     cy.contains('Scan barcodes')
