+++ tests/e2e/specs/reception/import_samples_from_sequencescape.cy.js
@@ -8,57 +8,7 @@ describe('Import samples from Sequencescape', () => {
     })
   })
 
-  // DEPRECATE-DPL-877 once enabled by default, remove this
-  it('Successfully - V1', () => {
-    cy.withFlags({
-      dpl_877_reception_request: { enabled: false },
-    })
-    cy.visit('#/reception')
-    cy.get('[data-type="source-list"]').select('Sequencescape')
-    cy.contains('Scan barcodes')
-    cy.get('#barcodes').type('DN9000002A\nNT1O')
-    cy.get('#cost_code').type('aCostCodeExample')
-    cy.get('[data-attribute=estimate_of_gb_required]').type('3')
-    cy.intercept(
-      {
-        url: '/api/v2/labware*',
-        query: {
-          'filter[barcode]': 'DN9000002A,NT1O',
-          include: 'receptacles.aliquots.sample.sample_metadata,receptacles.aliquots.study',
-          'fields[plates]': 'labware_barcode,receptacles',
-          'fields[tubes]': 'labware_barcode,receptacles',
-          'fields[aliquots]': 'study,library_type,sample',
-          'fields[receptacles]': 'aliquots',
-          'fields[sample_metadata]': 'sample_common_name',
-          'fields[samples]': 'sample_metadata,name,uuid',
-          'fields[studies]': 'uuid',
-          'fields[wells]': 'position,aliquots',
-        },
-      },
-      {
-        fixture: 'sequencescapeLabware.json',
-      },
-    )
-    cy.intercept('POST', '/v1/receptions', { fixture: 'tractionPacbioRequest.json' }).as(
-      'postPayload',
-    )
-
-    cy.contains('Import 2 labware into PacBio from Sequencescape')
-    cy.get('[data-action="import-labware"]').click()
-    // again, this should be 3 but 4 are in the response
-    cy.contains('Imported 4 labware(s) from Sequencescape')
-
-    cy.fixture('receptionCreateSourceSequencescape').then(({ data }) => {
-      cy.log('@postPayload')
-      cy.wait('@postPayload').its('request.body').should('deep.equal', data)
-    })
-  })
-
   it('Successfully - V2', () => {
-    // DEPRECATE-DPL-877 once enabled by default, remove this line
-    cy.withFlags({
-      dpl_877_reception_request: { enabled: true },
-    })
     cy.visit('#/reception')
     cy.get('[data-type="source-list"]').select('Sequencescape')
     cy.contains('Scan barcodes')
